Extract classes and functions from main file
The main file was pretty much the only file with code, and thus had both declarations and side-effects. This is a PSR-1 violation that causes many problems, such as https://github.com/inpsyde/idealo-for-woocommerce/pull/43#issuecomment-841413114 due to side-effects being triggered by autoloading.
This changeset extracts declarations out of the main plugin file and configures autoload paths to load them on demand. It also bumps the minimal required PHP version to 7.1: latest supported PHP version by WP is 5.6 anyway, it's a huge problem to configure an environment with PHP 5.2, and WP is [moving to PHP 7.1][trac-841413114] anyway soon.

[trac-841413114]: https://github.com/inpsyde/idealo-for-woocommerce/pull/43#issuecomment-841413114